backend/requirements.txt
Flask
psycopg2-binary~=2.8.6
flask_cors